Move LDAPSetting to the config module
Settings are configuration, it makes more sense for it to live there.
